Add tests for useMapCountryToColor
