+++ packages/app/jest.config.js
@@ -7,4 +7,7 @@ module.exports = {
   transform: {
     '^.+\\.(ts|tsx)$': 'ts-jest',
   },
+  moduleNameMapper: {
+    '~/(.*)': '<rootDir>/src/$1',
+  },
 };
