+++ packages/app/package.json
@@ -12,6 +12,7 @@
     "@sanity/client": "^2.11.0",
     "@styled-system/css": "^5.1.5",
     "@styled-system/should-forward-prop": "^5.1.5",
+    "@testing-library/react-hooks": "^7.0.2",
     "@tippyjs/react": "^4.2.5",
     "@turf/clean-coords": "^6.5.0",
     "@turf/helpers": "^6.5.0",
