+++ packages/icons/src/index.d.ts
@@ -103,9 +103,9 @@ export const Phone: Icon;
 export const PubliekToegankelijkeLocaties: Icon;
 export const Recreatie: Icon;
 export const Reizen: Icon;
+export const ReproductieGraf: Icon;
 export const ReproductiegetalSmall: Icon;
 export const Reproductiegetal: Icon;
-export const ReproductieGraf: Icon;
 export const Reset: Icon;
 export const RioolwaterMonitoring: Icon;
 export const SchoolAndDayCare: Icon;
@@ -113,10 +113,10 @@ export const SearchIconBold: Icon;
 export const SearchIcon: Icon;
 export const Search: Icon;
 export const SneezeCoughElbow: Icon;
-export const SporterMetZweetband: Icon;
 export const SportBinnensportlocaties: Icon;
 export const SportBuiten: Icon;
 export const SportWedstrijden: Icon;
+export const SporterMetZweetband: Icon;
 export const Stap1Avondklok: Icon;
 export const Stap1HorecaMax: Icon;
 export const Stap1HorecaPertafel: Icon;
