+++ packages/icons/src/index.js
@@ -94,9 +94,9 @@ export { default as Phone } from './icons/phone';
 export { default as PubliekToegankelijkeLocaties } from './icons/publiek-toegankelijke-locaties';
 export { default as Recreatie } from './icons/recreatie';
 export { default as Reizen } from './icons/reizen';
+export { default as ReproductieGraf } from './icons/reproductie_graf';
 export { default as ReproductiegetalSmall } from './icons/reproductiegetal-small';
 export { default as Reproductiegetal } from './icons/reproductiegetal';
-export { default as ReproductieGraf } from './icons/reproductie_graf';
 export { default as Reset } from './icons/reset';
 export { default as RioolwaterMonitoring } from './icons/rioolwater-monitoring';
 export { default as SchoolAndDayCare } from './icons/school_and_day_care';
@@ -104,10 +104,10 @@ export { default as SearchIconBold } from './icons/search-icon-bold';
 export { default as SearchIcon } from './icons/search-icon';
 export { default as Search } from './icons/search';
 export { default as SneezeCoughElbow } from './icons/sneeze_cough_elbow';
-export { default as SporterMetZweetband } from './icons/sporter-met-zweetband';
 export { default as SportBinnensportlocaties } from './icons/sport_binnensportlocaties';
 export { default as SportBuiten } from './icons/sport_buiten';
 export { default as SportWedstrijden } from './icons/sport_wedstrijden';
+export { default as SporterMetZweetband } from './icons/sporter-met-zweetband';
 export { default as Stap1Avondklok } from './icons/stap-1_avondklok';
 export { default as Stap1HorecaMax } from './icons/stap_1-horeca_max';
 export { default as Stap1HorecaPertafel } from './icons/stap_1-horeca_pertafel';
